requirements_local.txt
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
black==22.6.0
click==8.1.3
colorama==0.4.5
debugpy==1.6.0
decorator==5.1.1
dj-database-url==1.0.0
dj-static==0.0.6
Django==3.2.15
django-cors-headers==3.13.0
django-crispy-forms==1.13.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
entrypoints==0.4
executing==0.8.3
future==0.18.2
ipykernel==6.15.0
ipython==7.34.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
Markdown==3.4.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pywin32==304
pyzmq==23.2.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
static3==0.7.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.3.0
typing_extensions==4.1.1
tzdata==2022.1
wcwidth==0.2.5
